[refac] shape canonical axes
Use +X for normal/height/symmetry/extrusion axis. It will be easier to
place geometric shapes in the world, because components like Insertion
and Container uses the X axis to describe the direction of the insertion
or container mouth/opening.

For 2D shapes (except planes), the X axis direction shows how to make
contact with the shape.

For 3D shapes, the X axis describes the direction of the symmetry
starting at the bottom.